# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Visitor Management System API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/department/{department_id}/users": {
      "get": {
        "summary": "Get all users for a specific department",
        "tags": [
          "Departments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "department_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Department ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users for the department"
          }
        }
      }
    },
...
    "/api/v1/location/access-location": {
      "post": {
        "summary": "Record an access attempt for a location",
        "tags": [
          "Locations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "visitorId",
                  "locationId"
                ],
                "properties": {
                  "visitorId": {
                    "type": "integer"
                  },
                  "locationId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Access attempt recorded"
          }
        }
      }
    },
    "/api/v1/location": {
      "get": {
        "summary": "Get all locations",
        "tags": [
          "Locations"
        ],
        "responses": {
          "200": {
            "description": "List of locations"
          }
        }
      },
      "post": {
        "summary": "Create a new location",
        "tags": [
          "Locations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Location created successfully"
          }
        }
      }
    },
    "/api/v1/location/{location_id}": {
      "get": {
        "summary": "Get a location by ID",
        "tags": [
          "Locations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Location ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Location details"
          }
        }
      },
      "put": {
        "summary": "Update an existing location",
        "tags": [
          "Locations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Location ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a location",
        "tags": [
          "Locations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Location ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Location deleted successfully"
          }
        }
      }
    },
...
    "/api/v1/permission": {
      "get": {
        "summary": "Get all permissions",
        "tags": [
          "Permission"
        ],
        "responses": {
          "200": {
            "description": "List of permissions"
          }
        }
      },
      "post": {
        "summary": "Create a new permission",
        "tags": [
          "Permission"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "permission_name"
                ],
                "properties": {
                  "permission_name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permission created successfully"
          }
        }
      }
    },
    "/api/v1/permission/{permission_id}": {
      "get": {
        "summary": "Get a permission by ID",
        "tags": [
          "Permission"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "permission_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Permission ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Permission details"
          }
        }
      },
      "put": {
        "summary": "Update a permission",
        "tags": [
          "Permission"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "permission_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Permission ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permission_name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a permission",
        "tags": [
          "Permission"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "permission_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Permission ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Permission deleted successfully"
          }
        }
      }
    },
...
    "/api/v1/user/login": {
      "post": {
        "summary": "Login a user",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          }
        }
      }
    },
...
    "/api/v1/user/adminDashboard": {
      "get": {
        "summary": "Access admin dashboard (protected)",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "Admin dashboard info"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
...
    "/api/v1/user": {
      "get": {
        "summary": "Get all users",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "List of users"
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          }
        }
      }
    },
...
    "/api/v1/user/{user_id}": {
      "get": {
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details"
          }
        }
      },
      "put": {
        "summary": "Update a user",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          }
        }
      }
    },
...
  },
...
}